fix: refactor FullCalendar to fetch posts based on event ID and improve post handling
