+++ src/pages/Post.tsx
@@ -1,4 +1,4 @@
-import type { PostEventProps, Event } from '@/types/index';
+import type { Event } from '@/types/index';
 
 import { useState } from 'react';
 import {
@@ -13,7 +13,6 @@ import {
 import { LuMoreHorizontal } from 'react-icons/lu';
 import { useParams } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
-import { axiosAPIInstance } from '@/api/axios-config';
 
 import { FullCalendar } from '@/components/calendarComponents/fullCalendar';
 
@@ -44,19 +43,6 @@ export default function Post(props: Props) {
         ? 'Error fetching event'
         : 'Event not found';
 
-    const fetchPosts = async () => {
-        const response = await axiosAPIInstance.get(
-            `v1/event/${eventid}/posts`,
-        );
-
-        return response.data.data;
-    };
-
-    const { data: posts = [] } = useQuery<PostEventProps[]>({
-        queryKey: ['posts'],
-        queryFn: fetchPosts,
-    });
-
     const handleOpen = (backdrop: 'opaque' | 'transparent' | 'blur') => {
         setBackdrop(backdrop);
         onOpen();
@@ -66,7 +52,7 @@ export default function Post(props: Props) {
 
     const handleTabChange = (key: string) => {
         setActiveTab(key);
-    }
+    };
 
     return (
         <DefaultLayout>
@@ -110,7 +96,10 @@ export default function Post(props: Props) {
                         {props.children}
                     </Tab>
                     <Tab key="Calendar" title="Calendar">
-                        <FullCalendar posts={posts} onTabChange={handleTabChange} />
+                        <FullCalendar
+                            isPostType={false}
+                            onTabChange={handleTabChange}
+                        />
                     </Tab>
                     <Tab key="Member" title="Member">
                         <MembersPage />
